Added a way to test the building of examples from the command line.
This can ultimately be use for CI build testing.

This commit only supports build examples that do not depend on other
modules. It also only supports builds for AVR, but this can easily be
extended to all architectures that the Arduino-Makefile project
supports. See: https://github.com/sudar/Arduino-Makefile.git